+++ src/pixRequest/pixRequest.php
@@ -240,7 +240,7 @@ class PixRequest extends Resource
      */
     public static function parse($content, $signature, $user = null)
     {
-        $request = Parse::parseAndVerify($content, $signature, PixRequest ::resource(), $user);
+        $request = Parse::parseAndVerify($content, $signature, PixRequest::resource(), $user);
 
         $request->fee = $request->fee ? $request->fee : 0;
         $request->tags = $request->tags ? $request->tags : [];